Manual merge of estimation changes
Estimation using GRBL simulator (tweaked build with buffering disabled).
A lot of little things changed so that the file is preprocessed only
once. Alignment of command state updates to the UI are also fixed. The
estimation changes will only work for x86, so will need to be adjusted
before suitable for merge to master.