Fixed calculation for inch units for XY and XYZ probing.
Changed "Retract Amount" to mean only the retract distance between the fast and slow probes. Previously this was hard coded to 1 mm or 1 inch.

Changed the Z probe to return to the start position when finished (instead of using retract amount), to be consistent with the XY and XYZ probes.

A few other small changes to prepare for a later change to fix the problem of many error messages after an alarm (especially for XYZ probing).