Fix weird capabilities edge case.
With the capabilities merge in the GrblController raw response handler
it was possible for new capabilities to be added to an old version of
grbl. It would require quickly switching between a new version to an old
version.

This is no longer possible, also the tests are now more isolated.