Fixed problem with the console not sending commands on Windows  (#2130)
Fixed problem with the console not sending commands on Windows by reverting to the old CommandTextArea. 
Added the possibility to post multiple lines from the console by pasting multiline string.
Also enabled to send empty lines to the controller as this can be used by the user to determine if the controller is still online.