Added safe positioning when starting an autoleveler job (#2282)
When starting the autoleveler it will make sure to heighten the probe to the max probe height before moving to the XY start position. After the job is finished it will go to the max height and go back to start position.

Also fixed bug where if the controller was reporting coordinates in inches the autoleveling wouldn't work.